model: Move issue-related code from `ScanRecord` to `ScannerRun`
This prepares for removing `ScanRecord` and inlining its properties
into `ScannerRun`.
